Download Steam app list in run-time instead of compile-time (fix #3)
Issuing a GET request inside the initialization of Lazy proved more
difficult than expected - it's a non-async closure that ends up running
inside an async function. Using reqwest::blocking panics because it
attempts to create a new tokio runtime inside an already running one.

I considered several alternatives:
- Leave the code as-is, generating a json at build-time and using it to
  initialize the hashmap.
- Use a blocking HTTP library like ureq just for the initialization.
- Add an async init function and call it manually before creating the
  tokio runtime.
- Replace the global with a method that returns an initialized hashmap
  and pass it around.

But all these are bad. I ended up creating a thread where another tokio
runtime can be created, and using that to run the futures that fetch the
app list. This code is ugly but it's contained in the Lazy
initialization.

This can hopefully be improved one day when once_cell (or it's std
replacement) support async.